feat: working gateway
